+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Semere Gebreyesus 2024 rifabutin ([doi:10.1128/aac.00354-24](https://doi.org/10.1128/aac.00354-24)) -- HIV/TB co-infected children on lopinavir/ritonavir-based antiretroviral therapy.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.